v3.2.4 (cleanup): split EditGameDrawer + GamesTable, add TopOffline + AddedPerMonth charts, drop dead scripts
- Remove `scripts/discord_notify.py` (superseded by `poli0981/.github` reusable, no workflow ref) and `scripts/migrate_to_shards.py` (one-shot v2.2 migration, completed). Verified the other 8 "unreferenced" scripts all have live `bash/*.sh` wrappers and stay.
- Split `web/src/components/games/EditGameDrawer.tsx` (718 LOC) into 7 focused files under `edit/`: container keeps save handler + dialog shell; `EditFormFields`, `EditJsonView` (with `JsonDiff`), `EditChangePreview`, `ViewToggle` extract the four UI regions; `form-utils.ts` exports `gameToForm` + `formToPatch`; `types.ts` re-exports `FormState` + `View`.
- Split `web/src/components/games/GamesTable.tsx` (568 LOC) into container + 5 files under `table/`: container keeps the virtualizer (single-source row-height), filters, fuzzy search, sort, paging; `columns.tsx` extracts `COLS: ColDef[]` + width constants; `TableHeader`, `TableRow` (`React.memo`), `TableToolbar`, `ExportMenu` extract the visual regions. No behaviour change.
- Add `web/src/components/charts/TopOfflineBar.tsx` mirror of `TopOnlineBar` with offline-tier thresholds (20k/1.5k/300) matching `top_offline.py`. New page `/top-offline` registered in `App.tsx` + `Sidebar.tsx`.
- Add `web/src/components/charts/AddedPerMonthBar.tsx` aggregating `added_at` into `YYYY-MM` buckets; rendered as a third card on `charts/Time.tsx`.
- i18n: `nav.topOffline`, `topOffline.{title,subtitle,cardTitle}`, `charts.time.addedPerMonth` added to `en.json` + `vi.json`.
- README: drop duplicate "Quick Links" section; add `🎯 Top offline (md)` row; version badge `3.2.3` → `3.2.4`.
- Version bump `1.2.3` → `1.2.4` in `web/package.json`, `web/src-tauri/Cargo.toml` (+ Cargo.lock regen), `web/src-tauri/tauri.conf.json`.

Verification: `npm run typecheck`, `npm run build`, `cargo check` all green.
