+++ web/src/components/games/GamesTable.tsx
@@ -1,36 +1,15 @@
 import { useMemo, useRef, useState, useEffect } from "react";
 import { useVirtualizer } from "@tanstack/react-virtual";
-import { useTranslation } from "react-i18next";
 import Fuse from "fuse.js";
-import {
-  ArrowUp,
-  ArrowDown,
-  ExternalLink,
-  AlertTriangle,
-  ChevronLeft,
-  ChevronRight,
-  Download,
-  FileJson,
-  FileSpreadsheet,
-} from "lucide-react";
-import { Badge } from "../ui/badge";
-import { Button } from "../ui/button";
-import i18nDefault from "../../i18n";
-import { useFilters, PAGE_SIZES, type PageSize } from "../../stores/filters";
-import {
-  formatNumber,
-  parseIntSafe,
-  parseReleaseDate,
-  parseReviewPercent,
-} from "../../lib/utils";
+import { useFilters } from "../../stores/filters";
 import { extractAppid } from "../../lib/data-store";
-import { headerToCapsule } from "../../lib/image";
 import type { GameRecord } from "../../lib/schema";
-import { cn } from "../../lib/utils";
 import { GameDetailDrawer } from "./GameDetailDrawer";
 import { BulkActionBar } from "./BulkActionBar";
-import { exportCsv, exportJson } from "../../lib/export";
-import { recordIssues } from "../../lib/validation";
+import { COLS, TOTAL_WIDTH, SELECT_COL_WIDTH } from "./table/columns";
+import { TableHeader } from "./table/TableHeader";
+import { TableRow } from "./table/TableRow";
+import { TableToolbar } from "./table/TableToolbar";
 
 interface Props {
   records: GameRecord[];
@@ -38,207 +17,7 @@ interface Props {
   onRowOpen?: (game: GameRecord) => void;
 }
 
-interface ColDef {
-  key: string;
-  label: string;
-  width: number;
-  sortable?: boolean;
-  align?: "left" | "right" | "center";
-  render: (g: GameRecord) => React.ReactNode;
-  sortValue?: (g: GameRecord) => string | number;
-}
-
-const COLS: ColDef[] = [
-  {
-    key: "thumb",
-    label: "",
-    width: 56,
-    render: (g) =>
-      g.header_image ? (
-        <img
-          loading="lazy"
-          src={headerToCapsule(g.header_image)}
-          alt=""
-          width={92}
-          height={43}
-          className="h-8 w-16 rounded object-cover"
-          onError={(e) => {
-            const el = e.currentTarget;
-            if (el.src !== g.header_image) el.src = g.header_image;
-          }}
-        />
-      ) : (
-        <div className="h-8 w-16 rounded bg-muted" />
-      ),
-  },
-  {
-    key: "issues",
-    label: "",
-    width: 28,
-    render: (g) => {
-      const issues = recordIssues(g);
-      if (issues.length === 0) return null;
-      return (
-        <div
-          className="grid h-5 w-5 place-items-center rounded text-amber-400"
-          title={issues.map((i) => i.label).join(" · ")}
-        >
-          <AlertTriangle className="h-3 w-3" />
-        </div>
-      );
-    },
-  },
-  {
-    key: "name",
-    label: "Name",
-    width: 280,
-    sortable: true,
-    sortValue: (g) => g.name?.toLowerCase() ?? "",
-    render: (g) => (
-      <span className="font-medium">
-        {g.is_dead && <span title="Dead game (no players)" className="mr-1">💀</span>}
-        {g.name || "—"}
-      </span>
-    ),
-  },
-  {
-    key: "genre",
-    label: "Genre",
-    width: 120,
-    sortable: true,
-    sortValue: (g) => g.genre ?? "",
-    render: (g) =>
-      g.genre ? (
-        <Badge variant="outline" className="font-normal">
-          {g.genre}
-        </Badge>
-      ) : (
-        "—"
-      ),
-  },
-  {
-    key: "type_game",
-    label: "Type",
-    width: 80,
-    sortable: true,
-    sortValue: (g) => g.type_game ?? "",
-    render: (g) =>
-      g.type_game ? (
-        <Badge
-          variant={g.type_game === "online" ? "success" : "secondary"}
-          className="font-normal"
-        >
-          {g.type_game}
-        </Badge>
-      ) : (
-        "—"
-      ),
-  },
-  {
-    key: "reviews",
-    label: "Reviews",
-    width: 110,
-    sortable: true,
-    align: "right",
-    sortValue: (g) => parseReviewPercent(g.reviews) ?? -1,
-    render: (g) => {
-      const pct = parseReviewPercent(g.reviews);
-      if (pct === null) return <span className="text-muted-foreground">—</span>;
-      const color =
-        pct >= 80 ? "text-emerald-400" : pct >= 70 ? "text-amber-400" : "text-rose-400";
-      return <span className={cn("font-mono", color)}>{pct}%</span>;
-    },
-  },
-  {
-    key: "current_players",
-    label: "Players",
-    width: 110,
-    sortable: true,
-    align: "right",
-    sortValue: (g) => parseIntSafe(g.current_players),
-    render: (g) => (
-      <span className="font-mono">{formatNumber(g.current_players)}</span>
-    ),
-  },
-  {
-    key: "anti_cheat",
-    label: "AC",
-    width: 100,
-    sortable: true,
-    sortValue: (g) => g.anti_cheat ?? "",
-    render: (g) =>
-      g.anti_cheat && g.anti_cheat !== "-" ? (
-        <Badge
-          variant={g.is_kernel_ac ? "destructive" : "warning"}
-          className="font-normal"
-        >
-          {g.anti_cheat}
-        </Badge>
-      ) : (
-        <span className="text-muted-foreground">—</span>
-      ),
-  },
-  {
-    key: "platforms",
-    label: "Platforms",
-    width: 140,
-    render: (g) => (
-      <span className="text-xs text-muted-foreground">
-        {(g.platforms ?? []).join(", ") || "—"}
-      </span>
-    ),
-  },
-  {
-    key: "release_date",
-    label: "Released",
-    width: 120,
-    sortable: true,
-    // Parse Steam-style date strings ("22 Aug, 2012") to timestamps —
-    // string-compare sorts "22 Aug, 2012" < "3 Jan, 2025" alphabetically
-    // and would scramble years entirely.
-    sortValue: (g) => parseReleaseDate(g.release_date),
-    render: (g) => (
-      <span className="text-xs text-muted-foreground">{g.release_date || "—"}</span>
-    ),
-  },
-  {
-    key: "status",
-    label: "Status",
-    width: 90,
-    sortable: true,
-    sortValue: (g) => g.status ?? "",
-    render: (g) => (
-      <Badge
-        variant={g.status === "active" ? "secondary" : "destructive"}
-        className="font-normal"
-      >
-        {g.status}
-      </Badge>
-    ),
-  },
-  {
-    key: "link",
-    label: "",
-    width: 48,
-    render: (g) => (
-      <a
-        href={g.link}
-        target="_blank"
-        rel="noreferrer"
-        onClick={(e) => e.stopPropagation()}
-        className="text-muted-foreground hover:text-primary"
-        title={i18nDefault.t("edit.openOnSteam")}
-      >
-        <ExternalLink className="h-3.5 w-3.5" />
-      </a>
-    ),
-  },
-];
-
-const TOTAL_WIDTH = COLS.reduce((s, c) => s + c.width, 0);
-
 export function GamesTable({ records, onRowOpen }: Props) {
-  const { t } = useTranslation();
   const search = useFilters((s) => s.search);
   const genre = useFilters((s) => s.genre);
   const typeGame = useFilters((s) => s.type_game);
@@ -255,11 +34,9 @@ export function GamesTable({ records, onRowOpen }: Props) {
   const [detail, setDetail] = useState<GameRecord | null>(null);
   const [page, setPage] = useState(1);
 
-  // Clear selection on unmount.
   const clearSelection = useFilters((s) => s.clearSelection);
   useEffect(() => clearSelection, [clearSelection]);
 
-  // Reset page when filters or page size change.
   useEffect(() => {
     setPage(1);
   }, [search, genre, typeGame, platform, status, hideDead, pageSize]);
@@ -322,7 +99,6 @@ export function GamesTable({ records, onRowOpen }: Props) {
     overscan: 12,
   });
 
-  // Reset scroll to top whenever the page slice changes.
   useEffect(() => {
     containerRef.current?.scrollTo({ top: 0 });
   }, [safePage, pageSize, sortKey, sortDir]);
@@ -342,110 +118,33 @@ export function GamesTable({ records, onRowOpen }: Props) {
     return setSort(null, null);
   }
 
-  const SELECT_COL_WIDTH = 36;
   const totalWidth = TOTAL_WIDTH + SELECT_COL_WIDTH;
+  const handleOpen = onRowOpen ?? setDetail;
 
   return (
     <div className="space-y-3">
       <BulkActionBar visibleAppids={visibleAppids} />
 
       <div className="rounded-lg border bg-card">
-        <div className="flex flex-wrap items-center justify-between gap-2 border-b px-3 py-2 text-xs text-muted-foreground">
-          <span>
-            <strong className="text-foreground">{formatNumber(sorted.length)}</strong>{" "}
-            {t("games.ofTotalGames", { total: formatNumber(records.length) })}
-            {selected.size > 0 && (
-              <span className="ml-2 text-primary">
-                · {formatNumber(selected.size)} {t("common.selected")}
-              </span>
-            )}
-          </span>
-
-          <div className="flex flex-wrap items-center gap-1.5">
-            <ExportMenu records={sorted} />
-            <select
-              value={pageSize}
-              onChange={(e) => setPageSize(Number(e.target.value) as PageSize)}
-              className="h-7 rounded-md border bg-background px-1.5 text-xs"
-              title={t("system.rowsPerPageTitle")}
-            >
-              {PAGE_SIZES.map((n) => (
-                <option key={n} value={n}>
-                  {n === -1 ? t("common.all") : t("common.rowsPerPage", { n })}
-                </option>
-              ))}
-            </select>
-            {pageSize !== -1 && (
-              <>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  className="h-7 w-7 p-0"
-                  onClick={() => setPage((p) => Math.max(1, p - 1))}
-                  disabled={safePage <= 1}
-                  title={t("system.previousPage")}
-                >
-                  <ChevronLeft className="h-3.5 w-3.5" />
-                </Button>
-                <span className="px-1 font-mono">
-                  {safePage}/{totalPages}
-                </span>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  className="h-7 w-7 p-0"
-                  onClick={() => setPage((p) => Math.min(totalPages, p + 1))}
-                  disabled={safePage >= totalPages}
-                  title={t("system.nextPage")}
-                >
-                  <ChevronRight className="h-3.5 w-3.5" />
-                </Button>
-              </>
-            )}
-          </div>
-        </div>
+        <TableToolbar
+          totalRecords={records.length}
+          filteredCount={sorted.length}
+          selectedCount={selected.size}
+          sorted={sorted}
+          pageSize={pageSize}
+          setPageSize={setPageSize}
+          page={safePage}
+          totalPages={totalPages}
+          setPage={setPage}
+        />
 
         <div
           ref={containerRef}
           className="relative h-[calc(100vh-300px)] overflow-auto [&::-webkit-scrollbar]:hidden [scrollbar-width:none]"
         >
           <div style={{ width: totalWidth, minWidth: "100%" }}>
-            {/* Header */}
-            <div className="sticky top-0 z-10 flex border-b bg-card/95 backdrop-blur">
-              <div
-                className="flex h-10 items-center justify-center"
-                style={{ width: SELECT_COL_WIDTH, minWidth: SELECT_COL_WIDTH }}
-              />
-              {COLS.map((c) => {
-                const active = sortKey === c.key;
-                return (
-                  <div
-                    key={c.key}
-                    className={cn(
-                      "flex h-10 items-center px-3 text-xs font-semibold uppercase tracking-wider text-muted-foreground",
-                      c.align === "right" && "justify-end",
-                      c.align === "center" && "justify-center",
-                      c.sortable && "cursor-pointer select-none hover:text-foreground",
-                    )}
-                    style={{ width: c.width, minWidth: c.width }}
-                    onClick={() => c.sortable && toggleSort(c.key)}
-                  >
-                    <span>{c.label}</span>
-                    {c.sortable && active && (
-                      <span className="ml-1 inline-flex">
-                        {sortDir === "asc" ? (
-                          <ArrowUp className="h-3 w-3" />
-                        ) : (
-                          <ArrowDown className="h-3 w-3" />
-                        )}
-                      </span>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
+            <TableHeader sortKey={sortKey} sortDir={sortDir} toggleSort={toggleSort} />
 
-            {/* Virtual rows */}
             <div
               style={{
                 height: rowVirtualizer.getTotalSize(),
@@ -456,50 +155,16 @@ export function GamesTable({ records, onRowOpen }: Props) {
               {rowVirtualizer.getVirtualItems().map((vrow) => {
                 const g = paged[vrow.index];
                 const aid = extractAppid(g.link) ?? "";
-                const isSelected = selected.has(aid);
                 return (
-                  <div
+                  <TableRow
                     key={vrow.key}
-                    className={cn(
-                      "absolute left-0 top-0 flex w-full cursor-pointer items-center border-b border-border/50 hover:bg-accent/40",
-                      isSelected && "bg-primary/10 hover:bg-primary/15",
-                    )}
-                    style={{
-                      transform: `translateY(${vrow.start}px)`,
-                      height: vrow.size,
-                    }}
-                    onClick={() => (onRowOpen ? onRowOpen(g) : setDetail(g))}
-                    data-appid={aid}
-                  >
-                    <div
-                      className="flex items-center justify-center"
-                      style={{ width: SELECT_COL_WIDTH, minWidth: SELECT_COL_WIDTH }}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        if (aid) toggleSelect(aid);
-                      }}
-                    >
-                      <input
-                        type="checkbox"
-                        checked={isSelected}
-                        onChange={() => {}}
-                        className="h-3.5 w-3.5 rounded border-input bg-transparent accent-primary"
-                      />
-                    </div>
-                    {COLS.map((c) => (
-                      <div
-                        key={c.key}
-                        className={cn(
-                          "flex items-center px-3 py-2 text-sm",
-                          c.align === "right" && "justify-end",
-                          c.align === "center" && "justify-center",
-                        )}
-                        style={{ width: c.width, minWidth: c.width }}
-                      >
-                        {c.render(g)}
-                      </div>
-                    ))}
-                  </div>
+                    game={g}
+                    isSelected={selected.has(aid)}
+                    translateY={vrow.start}
+                    height={vrow.size}
+                    onSelect={toggleSelect}
+                    onOpen={handleOpen}
+                  />
                 );
               })}
             </div>
@@ -510,59 +175,6 @@ export function GamesTable({ records, onRowOpen }: Props) {
           <GameDetailDrawer game={detail} onClose={() => setDetail(null)} />
         )}
       </div>
-    </div>
-  );
-}
-
-function ExportMenu({ records }: { records: GameRecord[] }) {
-  const { t } = useTranslation();
-  const [open, setOpen] = useState(false);
-  const ts = new Date().toISOString().slice(0, 10);
-  return (
-    <div className="relative">
-      <Button
-        variant="ghost"
-        size="sm"
-        className="h-7 px-2 text-xs"
-        onClick={() => setOpen((o) => !o)}
-        title={t("games.exportTitle", { count: records.length })}
-      >
-        <Download className="mr-1 h-3 w-3" />
-        {t("games.exportLabel")}
-      </Button>
-      {open && (
-        <>
-          <div className="fixed inset-0 z-10" onClick={() => setOpen(false)} />
-          <div className="absolute right-0 top-full z-20 mt-1 w-44 rounded-md border bg-card p-1 shadow-lg">
-            <button
-              type="button"
-              onClick={() => {
-                exportCsv(records, `f2p-games-${ts}.csv`);
-                setOpen(false);
-              }}
-              className="flex w-full items-center gap-2 rounded px-2 py-1.5 text-left text-xs hover:bg-accent"
-            >
-              <FileSpreadsheet className="h-3 w-3" /> CSV
-              <span className="ml-auto text-muted-foreground">
-                {t("games.exportRowsCount", { count: records.length })}
-              </span>
-            </button>
-            <button
-              type="button"
-              onClick={() => {
-                exportJson(records, `f2p-games-${ts}.json`);
-                setOpen(false);
-              }}
-              className="flex w-full items-center gap-2 rounded px-2 py-1.5 text-left text-xs hover:bg-accent"
-            >
-              <FileJson className="h-3 w-3" /> JSON
-              <span className="ml-auto text-muted-foreground">
-                {t("games.exportRowsCount", { count: records.length })}
-              </span>
-            </button>
-          </div>
-        </>
-      )}
     </div>
   );
 }
